refactor word info popover with linegroup
